poprawione 2 i zrobione 3

+++ src/test/java/pl/javastart/RoznicaKwadratowTest.java
@@ -9,19 +9,19 @@ public class RoznicaKwadratowTest {
 
     @Test
     public void shouldRoznicaFor10() {
-        long wynik = roznicaKwadratow.roznicaKwadratow(10);
+        int wynik = roznicaKwadratow.roznicaKwadratow(10);
         Assert.assertThat(wynik, CoreMatchers.is(2640));
     }
 
     @Test
     public void shouldRoznicaFor2() {
-        long wynik = roznicaKwadratow.roznicaKwadratow(2);
+        int wynik = roznicaKwadratow.roznicaKwadratow(2);
         Assert.assertThat(wynik, CoreMatchers.is(4));
     }
 
     @Test
     public void shouldRoznicaFor5() {
-        long wynik = roznicaKwadratow.roznicaKwadratow(5);
+        int wynik = roznicaKwadratow.roznicaKwadratow(5);
         Assert.assertThat(wynik, CoreMatchers.is(170));
     }
 }
